+++ semgrep/tests/e2e/targets/basic/nosem.js
@@ -1,0 +1,9 @@
+nosem_func() // nosem
+nosem_func() // nosem: rules.test-nosem
+nosem_func() // NOSEM: rules.test-nosem
+
+nosem_func(
+    invalid_line // nosem: rules.test-nosem
+)
+
+nosem_func()